refactor : 역할에 맞게 패키지를 수정하다
