+++ src/main/java/nextstep/subway/applicaion/dto/response/SectionResponse.java
@@ -1,4 +1,4 @@
-package nextstep.subway.applicaion.dto;
+package nextstep.subway.applicaion.dto.response;
 
 public class SectionResponse {
