+++ src/main/java/nextstep/subway/domain/Repository/LineRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.Repository;
 
+import nextstep.subway.domain.Entity.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
