+++ src/main/java/nextstep/subway/domain/Repository/StationRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.Repository;
 
+import nextstep.subway.domain.Entity.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
