+++ src/test/java/nextstep/subway/step/SectionStep.java
@@ -3,7 +3,7 @@ package nextstep.subway.step;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import nextstep.subway.applicaion.dto.SectionRequest;
+import nextstep.subway.applicaion.dto.request.SectionRequest;
 import nextstep.subway.ui.path.SubwayPath;
 import org.springframework.http.MediaType;
